Laravel 5.2 Compatible
Removed two traits from your events in order to be compatible with Laravel 5.2: InteractsWithSockets and Dispatchable.

I was getting the same error as the issue #12 (https://github.com/antonioribeiro/google2fa-laravel/issues/12)

It seems to work on my app. Do you think this will impact the rest of the package?

Thanks!